+++ src/canvas/SelectableCanvas.ts
@@ -161,6 +161,9 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
   declare hoverCursor: CSSStyleDeclaration['cursor'];
   declare moveCursor: CSSStyleDeclaration['cursor'];
   declare defaultCursor: CSSStyleDeclaration['cursor'];
+  /**
+   * @deprecated use {@link BaseBrush#cursor} instead
+   */
   declare freeDrawingCursor: CSSStyleDeclaration['cursor'];
   declare notAllowedCursor: CSSStyleDeclaration['cursor'];
 
@@ -894,7 +897,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
    * );
    *
    */
-  getViewportPoint(e: TPointerEvent) {
+  getViewportPoint(e: TPointerEvent | DragEvent) {
     if (this._pointer) {
       return this._pointer;
     }
@@ -913,7 +916,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
    * );
    *
    */
-  getScenePoint(e: TPointerEvent) {
+  getScenePoint(e: TPointerEvent | DragEvent) {
     if (this._absolutePointer) {
       return this._absolutePointer;
     }
@@ -930,7 +933,7 @@ export class SelectableCanvas<EventSpec extends CanvasEvents = CanvasEvents>
    * @param {Boolean} [fromViewport] whether to return the point from the viewport or in the scene
    * @return {Point}
    */
-  getPointer(e: TPointerEvent, fromViewport = false): Point {
+  getPointer(e: TPointerEvent | DragEvent, fromViewport = false): Point {
     const upperCanvasEl = this.upperCanvasEl,
       bounds = upperCanvasEl.getBoundingClientRect();
     let pointer = getPointer(e),