+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- BREAKING(): Brush `onMouseDown` `onMouseMove` `onMouseUp` signature [#9501](https://github.com/fabricjs/fabric.js/pull/9501)
 - test(FabricObject): add a snapshot of the default values so that reordering and shuffling is verified. [#9492](https://github.com/fabricjs/fabric.js/pull/9492)
 - feat(FabricObject, Canvas) BREAKING: remove calculate true/false from the api. [#9483](https://github.com/fabricjs/fabric.js/pull/9483)
 - chore(): remove some Type assertions [#8950](https://github.com/fabricjs/fabric.js/pull/8950)